refactor(changes): extract BranchCommits into its own component
Pure move with no behavior change. Sets up CodeChanges.tsx to be a slim
orchestrator in subsequent tasks.
