+++ src/components/CodeChanges.tsx
@@ -1,12 +1,12 @@
 import { Component, createSignal, createEffect, on, Show, For, createMemo } from 'solid-js'
 import { createVirtualizer } from '@tanstack/solid-virtual'
-import { ChevronDown, ChevronRight, RefreshCw, X, GitCommit, Circle, GitCompare, FileText, ClipboardCopy, FolderOpen, ExternalLink, Tag } from 'lucide-solid'
+import { RefreshCw, X, GitCompare, FileText, ClipboardCopy, FolderOpen, ExternalLink, Tag } from 'lucide-solid'
 import { diffTabKey } from '../store/files'
 import { openDiffTab, openFilePinned, revealFileInTree, mainView, type DiffSource } from '../store/editorView'
 import { selectedTaskId } from '../store/ui'
 import { taskById } from '../store/tasks'
-import { clsx } from 'clsx'
 import { getFileIcon } from '../lib/fileIcons'
+import { BranchCommits } from './BranchCommits'
 import { taskGit, refreshTaskGit } from '../store/git'
 import * as ipc from '../lib/ipc'
 import type { GitStatus, FileStatus } from '../types'
@@ -41,16 +41,12 @@ export const CodeChanges: Component<Props> = (props) => {
   // null = uncommitted changes (read from store), string = commit hash (local)
   const [selectedCommit, setSelectedCommit] = createSignal<string | null>(null)
   const [commitStatus, setCommitStatus] = createSignal<GitStatus | null>(null)
-  const [commitsOpen, setCommitsOpen] = createSignal(
-    localStorage.getItem('verun:commitsOpen') !== 'false'
-  )
 
   // Reactive accessors that branch on selectedCommit
   const status = () => selectedCommit() ? commitStatus() : taskGit(props.taskId).status
   const commits = () => taskGit(props.taskId).commits
   const uncommittedCount = () => taskGit(props.taskId).status?.files.length ?? 0
   let fileScrollRef: HTMLDivElement | undefined
-  let commitScrollRef: HTMLDivElement | undefined
 
   const files = () => status()?.files || []
   const statsByPath = createMemo(() => {
@@ -117,33 +113,11 @@ export const CodeChanges: Component<Props> = (props) => {
     initialRect: { width: 280, height: 360 },
   })
 
-  const commitVirtualizer = createVirtualizer({
-    get count() { return commits().length },
-    getScrollElement: () => commitScrollRef ?? null,
-    estimateSize: () => 28,
-    overscan: 8,
-    initialRect: { width: 280, height: 192 },
-  })
-
   const visibleFileRows = () => {
     const rows = fileVirtualizer.getVirtualItems()
     if (rows.length > 0 || files().length === 0) return rows
     const size = 28
     return Array.from({ length: Math.min(files().length, 20) }, (_, index) => ({
-      key: index,
-      index,
-      start: index * size,
-      end: (index + 1) * size,
-      size,
-      lane: 0,
-    }))
-  }
-
-  const visibleCommitRows = () => {
-    const rows = commitVirtualizer.getVirtualItems()
-    if (rows.length > 0 || commits().length === 0) return rows
-    const size = 28
-    return Array.from({ length: Math.min(commits().length, 10) }, (_, index) => ({
       key: index,
       index,
       start: index * size,
@@ -181,23 +155,7 @@ export const CodeChanges: Component<Props> = (props) => {
     ]
   }
 
-  const toggleCommitsPanel = () => {
-    const next = !commitsOpen()
-    setCommitsOpen(next)
-    localStorage.setItem('verun:commitsOpen', String(next))
-  }
-
   const selectedCommitInfo = () => commits().find(c => c.hash === selectedCommit())
-
-  const formatTime = (ts: number) => {
-    const d = new Date(ts * 1000)
-    const now = Date.now()
-    const diff = now - d.getTime()
-    if (diff < 60_000) return 'just now'
-    if (diff < 3600_000) return `${Math.floor(diff / 60_000)}m ago`
-    if (diff < 86400_000) return `${Math.floor(diff / 3600_000)}h ago`
-    return `${Math.floor(diff / 86400_000)}d ago`
-  }
 
   return (
     <div class="flex flex-col h-full overflow-hidden min-w-0">
@@ -318,85 +276,12 @@ export const CodeChanges: Component<Props> = (props) => {
         </div>
       </div>
 
-      {/* Branch commits — collapsible, bottom-aligned like VS Code */}
-      <div class="h-px bg-outline/8 shrink-0" />
-      <div class="shrink-0">
-        <button
-          class="w-full h-8 flex items-center gap-1.5 px-3 text-xs hover:bg-surface-2"
-          onClick={toggleCommitsPanel}
-        >
-          {commitsOpen() ? <ChevronDown size={12} class="text-text-dim shrink-0" /> : <ChevronRight size={12} class="text-text-dim shrink-0" />}
-          <GitCommit size={12} class="text-text-dim shrink-0" />
-          <span class="font-medium text-text-secondary">Branch Commits</span>
-          <Show when={commits().length > 0}>
-            <span class="text-text-dim text-[10px] tabular-nums">
-              {commits().length}
-            </span>
-          </Show>
-        </button>
-
-        {/* Commits list */}
-        <Show when={commitsOpen()}>
-          <div ref={commitScrollRef} class="max-h-48 overflow-auto">
-            {/* Uncommitted changes tile */}
-            <button
-              class={clsx(
-                'relative w-full flex items-center gap-2 px-3 py-1.5 text-xs',
-                selectedCommit() === null
-                  ? 'bg-surface-2 text-text-primary'
-                  : 'hover:bg-surface-2 text-text-secondary',
-              )}
-              style={selectedCommit() === null ? { 'box-shadow': 'inset 2px 0 0 #2d6e4f' } : undefined}
-              onClick={() => selectCommit(null)}
-            >
-              <Circle size={11} class="shrink-0 text-text-dim" />
-              <span class="truncate flex-1 text-left">Uncommitted changes</span>
-              <Show when={uncommittedCount() > 0}>
-                <span class="text-[10px] text-text-dim shrink-0">{uncommittedCount()} file{uncommittedCount() !== 1 ? 's' : ''}</span>
-              </Show>
-            </button>
-
-            <div style={{ height: `${commitVirtualizer.getTotalSize()}px`, width: '100%', position: 'relative' }}>
-              <For each={visibleCommitRows()}>
-                {(vrow) => {
-                  const commit = () => commits()[vrow.index]
-                  return (
-                    <Show when={commit()}>
-                      {(c) => {
-                        const isSelected = () => selectedCommit() === c().hash
-                        return (
-                          <button
-                            class={clsx(
-                              'absolute left-0 top-0 w-full flex items-center gap-2 px-3 py-1.5 text-xs',
-                              isSelected() ? 'bg-surface-2 text-text-primary' : 'hover:bg-surface-2 text-text-secondary',
-                            )}
-                            style={{
-                              height: `${vrow.size}px`,
-                              transform: `translateY(${vrow.start}px)`,
-                              'box-shadow': isSelected() ? 'inset 2px 0 0 #2d6e4f' : undefined,
-                            }}
-                            onClick={() => selectCommit(c().hash)}
-                          >
-                            <span class="font-mono text-text-dim text-[10px] shrink-0">{c().shortHash}</span>
-                            <span class="truncate flex-1 text-left">{c().message}</span>
-                            <span class="text-[10px] text-text-dim shrink-0">{formatTime(c().timestamp)}</span>
-                          </button>
-                        )
-                      }}
-                    </Show>
-                  )
-                }}
-              </For>
-            </div>
-
-            <Show when={commits().length === 0 && !loading()}>
-              <div class="px-3 py-3 text-[11px] text-text-dim text-center">
-                No commits on this branch yet
-              </div>
-            </Show>
-          </div>
-        </Show>
-      </div>
+      <BranchCommits
+        taskId={props.taskId}
+        selectedCommit={selectedCommit()}
+        uncommittedCount={uncommittedCount()}
+        onSelectCommit={selectCommit}
+      />
 
       <ContextMenu
         open={!!fileMenu()}
